Changed the way we handle the tempFavorites state

+++ frontend/src/hooks/useHandleFavorites.js
@@ -1,41 +1,31 @@
 import { useState } from "react";
 import userService from "../services/userService";
 import { useDispatch } from "react-redux";
-import { addFavorite, removeFavorite } from "../redux/slices/tempFavoritesSlice";
+import { changeStateFavorite } from "../redux/slices/tempFavoritesSlice";
 const useHandleFavorites = (userId) => {
     const [loading, setLoading] = useState(false);
     const dispatch = useDispatch();
 
+    const updateFavorite = async (movieId, isFavorite) => {
+        setLoading(true);
+        try {
+            if (isFavorite) {
+                await userService.addMovieToFavorites(userId, movieId);
+            } else {
+                await userService.removeMovieFromFavorites(userId, movieId);
+            }
+            dispatch(changeStateFavorite({ movieId, isFavorite }));
+            return true;
+        } catch (error) {
+            console.error(error);
+            setLoading(false);
+            return false;
+        }
+        
+    };
     
-    const addMovieToFavorites = async (movieId) => {
-        setLoading(true);
-        try {
-            await userService.addMovieToFavorites(userId, movieId);
-            dispatch(addFavorite(movieId));
-            return true
-        } catch (error) {
-            console.log(error);
-            return false;
-        } finally {
-            setLoading(false);
-        }
-    };
 
-    const removeMovieFromFavorites = async (movieId) => {
-        setLoading(true);
-        try {
-            await userService.removeMovieFromFavorites(userId, movieId);
-            dispatch(removeFavorite(movieId));
-            return true
-        } catch (error) {
-            console.log(error);
-            return false;
-        } finally {
-            setLoading(false);
-        }
-    };
-
-    return { addMovieToFavorites, removeMovieFromFavorites, loading };
+    return { updateFavorite, loading };
 };
 
 export default useHandleFavorites;
